Performance: Take LOD into account for objects with a single LOD level.
Previously, these objects would always be drawn because it was assumed
they might not have a LOD defined at all. This meant that we weren't
taking LOD into account for objects that _did_ define a LOD, but only a
single level.

In addition, test for LOD early in draw_obj_at_ll() and draw_obj_at_xyz()
to avoid having to do all the OpenGL state setup in the case where we
aren't going to draw a LOD level at all anyway.