Reduce minimum line thickness in preview window.
The minimum line thickness was previously 5 pixels, which is appropriate
for the map window because below that thickness, it draws a "structural
preview". However, in the preview window, we draw nothing, and you had
to be pretty close to the line to see it. The minimum line thickness is
now one pixel, which yields more satisfactory results.