Make polygon texture fixed relative to the object rather than the window.
We make this change in behavior because in the 3D preview window,
there is no window origin that the texture could be fixed relative
to.

I would argue that this looks better even in the map window because the
texture stays fixed relative to the object (which is what we want
conceptually) and doesn't move around when the viewport changes.