Make WED_PreviewLayer ready to handle 3D preview.
At a high level, this means turning WED_PreviewLayer from a
WED_MapLayerWithZoomer into just a WED_MapLayer so that we can pass a
WED_PerspectiveCamera to it.

A consequence of this is that we have to modify a whole host of
lower-level functions so that they can take a WED_MapProjection and/or
WED_Camera instead of a WED_MapZoomerNew.

It would be pretty laborious to do this incrementally, so this commit
changes these functions in a single step; unfortunately, this does make
the commit relatively large.