Fix z-fighting on ATTR_draped parts of objects.
To do this, we draw the ATTR_draped parts of objects without z-buffering,
then later draw the remaining parts of the objects with z-buffering.
We treat tiles in .agps the same way as ATTR_draped parts of objects.

We don't draw strings or facades in two phases like this, even though
they could in principle also contain objects with ATTR_draped parts. In
practice, this doesn't seem to happen, so we'll deal with it once it
becomes a problem.